removed authorization where it made no sense; added bid endpoint

+++ backend/controllers/auction.js
@@ -1,4 +1,5 @@
 const Auction = require('../models/auction');
+const Order = require('../models/order');
 const {errorHandler} = require('../helpers/dbErrorHandler');
 
 exports.auctionById = (req, res, next, id) => {
@@ -25,7 +26,7 @@ exports.list = (req, res) => {
 
 exports.create = (req, res) => {
     req.body.imagePath = req.file.path;
-    req.body.owner = req.params.userId;
+    req.body.owner = req.auth._id;
     const auction = new Auction(req.body);
     auction.save((err, auction) => {
         if (err) return errorHandler(res, err);
@@ -45,8 +46,29 @@ exports.update = (req, res) => {
 };
 
 exports.remove = (req, res) => {
-    Auction.findByIdAndDelete(req.auction._id, (err, auction) => {
-        if (err) return errorHandler(res, err);
-        res.json(auction);
+    if (Order.exists({auction: req.auction})) {
+        return res.status(400).json({
+            message: 'Can not delete. There are orders using this auction!'
+        });
+    } else {
+        Auction.findByIdAndDelete(req.auction._id, (err, auction) => {
+            if (err) return errorHandler(res, err);
+            res.json(auction);
+        });
+    }
+};
+
+exports.bid = (req, res) => {
+    Auction.findOneAndUpdate(
+        {_id: req.auction._id, buyer: {$ne: req.auth._id}, price: {$st: req.body.price}},
+        { $set: { status: req.body.price, buyer: req.auth._id } },
+        (err, order) => {
+        if (err) {
+            return res.status(400).json({
+                error: errorHandler(err)
+            });
+        }
+        if (!order) return res.json({error: 'Bid not successful'});
+        res.json(order)
     });
 };
